# Conflicts:
#	drivers/dma/dma_stm32.c
#	drivers/gpio/gpio_infineon.c
#	drivers/i2c/i2c_infineon_pdl.c
#	drivers/modem/modem_cellular.c
#	drivers/pwm/pwm_capture.c
#	drivers/serial/uart_infineon_pdl.c
#	drivers/serial/uart_mcux_lpuart.c
#	subsys/dfu/img_util/flash_img.c
#	subsys/mgmt/mcumgr/transport/src/smp_udp.c
#	subsys/usb/device_next/class/usbd_cdc_acm.c
#	west.yml

+++ subsys/kvss/zms/zms.c
@@ -2366,7 +2366,7 @@ static int zms_iter_filter_common(struct zms_fs *fs, struct zms_iter *iter,
 	if (rc) {
 		return rc;
 	}
-	*previous_sector_num = SECTOR_NUM(ate_addr);
+	*previous_sector_num = (int)SECTOR_NUM(ate_addr);
 
 	if (!zms_ate_valid_different_sector(fs, ate, *current_cycle)) {
 		return 0;
